#132 Further clean up server API
- Move  classes from root package into a matching subpackage
- Create "org.eclipse.glsp.server.features.core.model" package that contains all actions and related classes that are needed for model loading
- Create a dedicated feature subpackage for undo/redo

Resolves eclipse-glsp/glsp/issues/132
