+++ plugins/org.eclipse.glsp.server/META-INF/MANIFEST.MF
@@ -16,19 +16,21 @@ Require-Bundle: org.apache.commons.io;bundle-version="2.2.0",
  javax.websocket;bundle-version="1.0.0";visibility:=reexport,
  org.eclipse.glsp.graph;bundle-version="[0.8.0,1.0.0)";visibility:=reexport
 Import-Package: com.google.inject.multibindings;version="1.3.0"
-Export-Package: org.eclipse.glsp.server,
- org.eclipse.glsp.server.actions,
+Export-Package: org.eclipse.glsp.server.actions,
+ org.eclipse.glsp.server.di,
  org.eclipse.glsp.server.diagram,
- org.eclipse.glsp.server.factory,
+ org.eclipse.glsp.server.disposable,
  org.eclipse.glsp.server.features.clipboard,
  org.eclipse.glsp.server.features.commandpalette,
  org.eclipse.glsp.server.features.contextactions,
  org.eclipse.glsp.server.features.contextmenu,
+ org.eclipse.glsp.server.features.core.model,
  org.eclipse.glsp.server.features.directediting,
  org.eclipse.glsp.server.features.modelsourcewatcher,
  org.eclipse.glsp.server.features.navigation,
  org.eclipse.glsp.server.features.popup,
  org.eclipse.glsp.server.features.toolpalette,
+ org.eclipse.glsp.server.features.undoredo,
  org.eclipse.glsp.server.features.validation,
  org.eclipse.glsp.server.internal.action;x-internal:=true,
  org.eclipse.glsp.server.internal.di;x-internal:=true,
