+++ plugins/org.eclipse.glsp.server/src/org/eclipse/glsp/server/actions/ActionHandler.java
@@ -21,8 +21,8 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
-import org.eclipse.glsp.server.Handler;
 import org.eclipse.glsp.server.model.GModelState;
+import org.eclipse.glsp.server.utils.Handler;
 
 public interface ActionHandler extends Handler<Action> {
    List<Class<? extends Action>> getHandledActionTypes();
