+++ plugins/org.eclipse.glsp.server/src/org/eclipse/glsp/server/features/core/model/RequestModelAction.java
@@ -13,9 +13,11 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ******************************************************************************/
-package org.eclipse.glsp.server.actions;
+package org.eclipse.glsp.server.features.core.model;
 
 import java.util.Map;
+
+import org.eclipse.glsp.server.actions.Action;
 
 public class RequestModelAction extends Action {
