+++ plugins/org.eclipse.glsp.server/src/org/eclipse/glsp/server/features/undoredo/RedoAction.java
@@ -13,12 +13,16 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ******************************************************************************/
-package org.eclipse.glsp.server;
+package org.eclipse.glsp.server.features.undoredo;
 
-public interface Handler<T> {
+import org.eclipse.glsp.server.actions.Action;
 
-   default int getPriority() { return Integer.MIN_VALUE; }
+public class RedoAction extends Action {
 
-   boolean handles(T object);
+   public static final String ID = "glspRedo";
+
+   public RedoAction() {
+      super(ID);
+   }
 
 }
