+++ plugins/org.eclipse.glsp.server/src/org/eclipse/glsp/server/internal/di/MapMultiRegistry.java
@@ -22,8 +22,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import org.eclipse.glsp.server.MultiRegistry;
 import org.eclipse.glsp.server.protocol.GLSPServerException;
+import org.eclipse.glsp.server.utils.MultiRegistry;
 
 public abstract class MapMultiRegistry<K, V> implements MultiRegistry<K, V> {
    protected Map<K, List<V>> elements;
