+++ plugins/org.eclipse.glsp.server/src/org/eclipse/glsp/server/internal/di/MultiBindingDefaults.java
@@ -21,35 +21,34 @@ import org.eclipse.glsp.server.actions.Action;
 import org.eclipse.glsp.server.actions.ActionHandler;
 import org.eclipse.glsp.server.actions.CenterAction;
 import org.eclipse.glsp.server.actions.ClientActionHandler;
-import org.eclipse.glsp.server.actions.ComputedBoundsActionHandler;
 import org.eclipse.glsp.server.actions.ConfigureServerHandlersAction;
 import org.eclipse.glsp.server.actions.DisposeClientSessionActionHandler;
 import org.eclipse.glsp.server.actions.ExportSVGAction;
 import org.eclipse.glsp.server.actions.FitToScreenAction;
 import org.eclipse.glsp.server.actions.GLSPServerStatusAction;
 import org.eclipse.glsp.server.actions.InitializeClientSessionActionHandler;
-import org.eclipse.glsp.server.actions.RequestBoundsAction;
-import org.eclipse.glsp.server.actions.RequestModelActionHandler;
 import org.eclipse.glsp.server.actions.SaveModelActionHandler;
 import org.eclipse.glsp.server.actions.SelectAction;
 import org.eclipse.glsp.server.actions.SelectAllAction;
 import org.eclipse.glsp.server.actions.ServerMessageAction;
 import org.eclipse.glsp.server.actions.ServerStatusAction;
-import org.eclipse.glsp.server.actions.SetBoundsAction;
 import org.eclipse.glsp.server.actions.SetDirtyStateAction;
 import org.eclipse.glsp.server.actions.SetEditModeAction;
 import org.eclipse.glsp.server.actions.SetEditModeActionHandler;
-import org.eclipse.glsp.server.actions.SetModelAction;
 import org.eclipse.glsp.server.actions.TriggerEdgeCreationAction;
 import org.eclipse.glsp.server.actions.TriggerNodeCreationAction;
-import org.eclipse.glsp.server.actions.UndoRedoActionHandler;
-import org.eclipse.glsp.server.actions.UpdateModelAction;
 import org.eclipse.glsp.server.diagram.RequestTypeHintsActionHandler;
 import org.eclipse.glsp.server.diagram.SetTypeHintsAction;
 import org.eclipse.glsp.server.features.clipboard.RequestClipboardDataActionHandler;
 import org.eclipse.glsp.server.features.clipboard.SetClipboardDataAction;
 import org.eclipse.glsp.server.features.contextactions.RequestContextActionsHandler;
 import org.eclipse.glsp.server.features.contextactions.SetContextActions;
+import org.eclipse.glsp.server.features.core.model.ComputedBoundsActionHandler;
+import org.eclipse.glsp.server.features.core.model.RequestBoundsAction;
+import org.eclipse.glsp.server.features.core.model.RequestModelActionHandler;
+import org.eclipse.glsp.server.features.core.model.SetBoundsAction;
+import org.eclipse.glsp.server.features.core.model.SetModelAction;
+import org.eclipse.glsp.server.features.core.model.UpdateModelAction;
 import org.eclipse.glsp.server.features.directediting.ApplyLabelEditOperationHandler;
 import org.eclipse.glsp.server.features.directediting.RequestEditValidationHandler;
 import org.eclipse.glsp.server.features.directediting.SetEditValidationResultAction;
@@ -61,6 +60,7 @@ import org.eclipse.glsp.server.features.navigation.SetNavigationTargetsAction;
 import org.eclipse.glsp.server.features.navigation.SetResolvedNavigationTargetAction;
 import org.eclipse.glsp.server.features.popup.RequestPopupModelActionHandler;
 import org.eclipse.glsp.server.features.popup.SetPopupModelAction;
+import org.eclipse.glsp.server.features.undoredo.UndoRedoActionHandler;
 import org.eclipse.glsp.server.features.validation.DeleteMarkersAction;
 import org.eclipse.glsp.server.features.validation.RequestMarkersHandler;
 import org.eclipse.glsp.server.features.validation.SetMarkersAction;
