+++ plugins/org.eclipse.glsp.server/src/org/eclipse/glsp/server/launch/GLSPServerLauncher.java
@@ -19,7 +19,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import org.eclipse.glsp.server.GLSPModule;
+import org.eclipse.glsp.server.di.GLSPModule;
 
 import com.google.inject.Guice;
 import com.google.inject.Injector;
